requirements.txt
pymongo
requests
turbopuffer
tqdm
pyyaml
